requirements.txt
accelerate==0.32.1
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
contractions==0.1.73
cycler==0.12.1
datasets==2.20.0
dill==0.3.8
emoji==2.12.1
evaluate==0.4.2
filelock==3.14.0
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.5.0
gensim==4.3.2
huggingface-hub==0.23.4
hydra-core==1.3.2
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
nltk==3.8.1
numpy==1.26.4
omegaconf==2.3.0
packaging==24.0
pandas==2.2.2
pillow==10.2.0
psutil==6.0.0
pyahocorasick==2.1.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.10.1
six==1.16.0
smart-open==7.0.4
sympy==1.12.1
tbb==2021.12.0
textsearch==0.0.24
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.66.4
transformers==4.44.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.2
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
